Move `ssl_check.py` to `chia/ssl`
